requirements.txt
flask==2.2.3
python-dotenv==1.0.0
langchain==0.0.235
openai==0.28.0
pinecone-client==2.2.1
tiktoken==0.4.0
chromadb==0.4.6
pydantic==1.10.8